requirements.txt
pillow
aiohttp
aiounittest